requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
torch==2.9.1
openai-whisper==20250625
requests==2.32.5
redis==7.1.0
rq==2.6.1
slowapi==0.1.6
